genAI_book/FashionCompliment/backend/requirements.txt
Flask==2.3.2
gunicorn==22.0.0
google-cloud-aiplatform==1.111.0
pillow==11.3.0
